Fix possible nullpointer exception

+++ src/test/java/com/exasol/adapter/document/files/extension/ExtensionIT.java
@@ -55,8 +55,10 @@ class ExtensionIT {
 
     @AfterAll
     static void teardown() {
-        s3TestSetup.emptyS3Bucket(s3BucketName);
-        s3TestSetup.deleteBucket(s3BucketName);
+        if (s3TestSetup != null) {
+            s3TestSetup.emptyS3Bucket(s3BucketName);
+            s3TestSetup.deleteBucket(s3BucketName);
+        }
         if (setup != null) {
             setup.close();
         }
